# Specialize `WithPosition::fold`