Each call to `f` is done with a `Position` variant known at compile time so I think it might be optimized away in some cases?

    Without specialization: [5.3231 µs 5.3487 µs 5.3806 µs]
    With specialization:    [2.6974 µs 2.7042 µs 2.7132 µs]

We should mention #755 in PRs that specialize `fold` (such as this one) to help keep track of the advancement of the TODO list I'm gonna make there.