@Philippe-Cholet 2023-10-03
@phimuemue 
I do not update `handled_first` field anymore, as it's consumed.
I avoid to `peek` entirely, the bench become even faster.
And rebased on master.

    [5.4031 µs 5.4195 µs 5.4353 µs] // without specialization
    [563.64 ns 565.92 ns 568.61 ns] // with new specialization