@Philippe-Cholet 2023-10-04
Just adding that I initially tried to avoid `std::mem::swap` with `(init, head) = self.peekable.fold((init, head), ...);` which worked for me but not for the msrv CI test. It did not like assigning to a tuple. The performance was not affected.